SERVICES-709 remove ConfigModule.forRoot statement; rarity rank float-> long mapping

+++ src/modules/nft-rarity/nft-rarity.service.ts
@@ -543,7 +543,7 @@ export class NftRarityService {
               },
               {
                 key: 'nft_rarity_rank',
-                value: 'float',
+                value: 'long',
               },
             ]),
           );
